include singleton filter
